compress_service/requirements.txt
pillow~=10.4.0
watchdog~=4.0.1
tabulate~=0.9.0
boto3==1.34.158